moved no-longer-hook functions to utils directory
